Separate the artwork type
