+++ src/app/_types/Artwork.tsx
@@ -1,0 +1,13 @@
+type Artwork = {
+    type: "artwork" | null;
+    title: string;
+    description: string;
+    link: string;
+    img1: string;
+    img2?: string;
+    alt?: string;
+    github?: string;
+    category: string;
+};
+
+export default Artwork;
